+++ .flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"path_provider","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider-2.0.2\\\\","dependencies":[]},{"name":"sqflite","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\sqflite-2.0.0+3\\\\","dependencies":[]},{"name":"sqlite3_flutter_libs","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\sqlite3_flutter_libs-0.5.0\\\\","dependencies":[]}],"android":[{"name":"path_provider","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider-2.0.2\\\\","dependencies":[]},{"name":"sqflite","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\sqflite-2.0.0+3\\\\","dependencies":[]},{"name":"sqlite3_flutter_libs","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\sqlite3_flutter_libs-0.5.0\\\\","dependencies":[]}],"macos":[{"name":"path_provider_macos","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_macos-2.0.0\\\\","dependencies":[]},{"name":"sqflite","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\sqflite-2.0.0+3\\\\","dependencies":[]},{"name":"sqlite3_flutter_libs","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\sqlite3_flutter_libs-0.5.0\\\\","dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_linux-2.0.0\\\\","dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_windows-2.0.1\\\\","dependencies":[]}],"web":[]},"dependencyGraph":[{"name":"path_provider","dependencies":["path_provider_macos","path_provider_linux","path_provider_windows"]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_macos","dependencies":[]},{"name":"path_provider_windows","dependencies":[]},{"name":"sqflite","dependencies":[]},{"name":"sqlite3_flutter_libs","dependencies":[]}],"date_created":"2021-06-18 13:56:55.372871","version":"2.2.1"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"path_provider","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider-2.0.2\\\\","dependencies":[]},{"name":"sqflite","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\sqflite-2.0.0+3\\\\","dependencies":[]},{"name":"sqlite3_flutter_libs","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\sqlite3_flutter_libs-0.5.0\\\\","dependencies":[]}],"android":[{"name":"path_provider","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider-2.0.2\\\\","dependencies":[]},{"name":"sqflite","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\sqflite-2.0.0+3\\\\","dependencies":[]},{"name":"sqlite3_flutter_libs","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\sqlite3_flutter_libs-0.5.0\\\\","dependencies":[]}],"macos":[{"name":"path_provider_macos","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_macos-2.0.0\\\\","dependencies":[]},{"name":"sqflite","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\sqflite-2.0.0+3\\\\","dependencies":[]},{"name":"sqlite3_flutter_libs","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\sqlite3_flutter_libs-0.5.0\\\\","dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_linux-2.0.0\\\\","dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"E:\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_windows-2.0.1\\\\","dependencies":[]}],"web":[]},"dependencyGraph":[{"name":"path_provider","dependencies":["path_provider_macos","path_provider_linux","path_provider_windows"]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_macos","dependencies":[]},{"name":"path_provider_windows","dependencies":[]},{"name":"sqflite","dependencies":[]},{"name":"sqlite3_flutter_libs","dependencies":[]}],"date_created":"2021-06-21 08:56:09.125675","version":"2.2.1"}
